(chore) Rename app to api
