+++ blst/settings.py
@@ -41,7 +41,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'rest_framework',
     'rest_framework_swagger',
-    'app',
+    'api',
 ]
 
 MIDDLEWARE_CLASSES = [
